[Refactor] [FE-12] Content의 Time과 Location을 각 함수로 분리
